b/restate/core/serializers.py
from rest_framework import serializers
from .models import Property, Media

class MediaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Media
        fields = "__all__"

class PropertySerializer(serializers.ModelSerializer):
    media = MediaSerializer(many=True, read_only=True)

    class Meta:
        model = Property
        fields = "__all__"