b/restate/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
dj-database-url==2.2.0
pymongo==4.8.0
gunicorn==22.0.0
whitenoise==6.7.0
brotli==1.1.0